Small changes on the human mesh tests...

+++ src/hash/hash_common.h
@@ -8,7 +8,7 @@
 #include <stdbool.h>
 
 struct point_3d {
-    double x, y, z;
+    float x, y, z;
 };
 
 bool point_equals(struct point_3d a, struct point_3d b);
